backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.0.0
cos-python-sdk-v5==1.9.25
cryptography>=41.0.0